+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_infos/config.hpp
@@ -1,0 +1,9 @@
+class cfgWeapons
+{
+	#include "weapons\config.hpp"
+};
+
+class cfgVehicles
+{
+	#include "vehicles\config.hpp"
+};
