+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/weapons/armour_standard/config.hpp
@@ -1,0 +1,2 @@
+#include "odst\config.hpp"
+#include "marines2\config.hpp"
