+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/weapons/config.hpp
@@ -1,0 +1,1 @@
+#include "armour_standard\config.hpp"
